Move BattleDisplay.java to its own class and update method vis as needed
